about to add functionality
the service for multiple components must be made injectable

Services that are injectable and used by a component
become dependencies of that component. The component depends
on those services and can't function without them.

Dependency injection is the mechanism that manages the dependencies
of an app's components and the services that other components can use.